the method specified in the description has been moved to the shape class, custom methods that have been redefined have been removed
